rye-lilux/lilux/pyproject.toml
[project]
name = "lilux"
version = "0.1.0"
description = "Lilux MCP Kernel - 5 primitive tools for AI agent orchestration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["mcp", "ai", "agents", "kernel", "primitives"]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "packaging>=24.0",
    "jsonschema>=4.0.0",
    "keyring>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
rag = [
    "numpy>=1.20.0",
]

[project.scripts]
lilux = "lilux.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lilux*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
